chore: Update icons
