--- src/Playroom/ShareMenu/ChevronIcon.tsx
@@ -1,14 +1,0 @@
-import React from 'react';
-
-export default () => (
-  <svg
-    viewBox="0 0 24 24"
-    focusable="false"
-    fill="currentColor"
-    width={16}
-    height={16}
-    role="img"
-  >
-    <path d="M20.7 7.3c-.4-.4-1-.4-1.4 0L12 14.6 4.7 7.3c-.4-.4-1-.4-1.4 0s-.4 1 0 1.4l8 8c.2.2.5.3.7.3s.5-.1.7-.3l8-8c.4-.4.4-1 0-1.4z" />
-  </svg>
-);
