+++ src/Playroom/Toolbar/icons/ChevronSvg.tsx
@@ -6,12 +6,17 @@ import styles from './ChevronSvg.less';
 
 interface ChevronSvgProps extends SVGProps<SVGSVGElement> {
   direction?: 'down' | 'up' | 'left' | 'right';
+  size?: number;
 }
 
-export default ({ direction = 'down', ...props }: ChevronSvgProps) => (
+export default ({
+  direction = 'down',
+  size = 24,
+  ...props
+}: ChevronSvgProps) => (
   <svg
-    width="24"
-    height="24"
+    width={size}
+    height={size}
     viewBox="0 0 24 24"
     focusable="false"
     fill="currentColor"
